refactor(api-compare): resolve the moved interface declaration names
The RFC 0080 kind exemption deliberately keeps `moved` interface
declaration names scored — 22 of them, never audited. Resolve each:

- `Quoting` was a misplaced port: `abstract/quoting-interface.ts` declared
  the contract `ConnectionAdapters::Quoting` defines in `quoting.rb`.
  Moved into `abstract/quoting.ts`, its Rails-layout file, where 15 of its
  16 extras are now allowed by the Ruby file. The one left,
  `quoteIdentifier`, is a trails-only synonym of `quoteColumnName` and
  carries a PERMANENT tag.
- 16 are coincidental name collisions (duck-type shapes Ruby leaves
  unnamed, sharing a name with an unrelated Rails constant) — tagged
  PERMANENT naming the collision.
- 5 stand in for surface not yet ported — tagged CONVERGEABLE, each
  naming a story registered under RFC 0023.

Extras 4221 -> 4150.
