+++ packages/activerecord/src/fixtures.ts
@@ -15,7 +15,7 @@ import {
 import type { AbstractAdapter as DatabaseAdapter } from "./connection-adapters/abstract-adapter.js";
 import { Base } from "./base.js";
 import { findStiClass } from "./inheritance.js";
-import type { Quoting } from "./connection-adapters/abstract/quoting-interface.js";
+import type { Quoting } from "./connection-adapters/abstract/quoting.js";
 import { currentTimeFromProperTimezone } from "./timestamp.js";
 import { singularize, underscore } from "@blazetrails/activesupport";
 import { EncryptedAttributeType } from "./encryption/encrypted-attribute-type.js";
